Alternate implementation of #20 that adds [I]BitmexMultiAccountApiService
