+++ Bitmex.NET/IBitmexMultiAccountApiService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace Bitmex.NET
+{
+    public interface IBitmexMultiAccountApiService : IDisposable
+    {
+        Task<TResult> Execute<TParams, TResult>(IBitmexAuthorization bitmexAuthorization, ApiActionAttributes<TParams, TResult> apiAction,
+            TParams @params);
+    }
+}
